Implement 16-byte vertex format

+++ src/main/java/me/jellysquid/mods/sodium/client/render/chunk/terrain/material/parameters/MaterialParameters.java
@@ -4,8 +4,8 @@ public class MaterialParameters {
     public static final int OFFSET_USE_MIP = 0;
     public static final int OFFSET_ALPHA_CUTOFF = 1;
 
-    public static byte pack(AlphaCutoffParameter alphaCutoff, boolean useMipmaps) {
-        return (byte) (((useMipmaps ? 1 : 0) << OFFSET_USE_MIP) |
-                        ((alphaCutoff.ordinal()) << OFFSET_ALPHA_CUTOFF));
+    public static int pack(AlphaCutoffParameter alphaCutoff, boolean useMipmaps) {
+        return (((useMipmaps ? 1 : 0) << OFFSET_USE_MIP) |
+                ((alphaCutoff.ordinal()) << OFFSET_ALPHA_CUTOFF));
     }
 }
